backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
drf-yasg==1.21.7
python-dotenv==1.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
mysqlclient==2.2.0
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
django-filter==23.3
django-storages==1.14.2
boto3==1.28.70
whitenoise==6.6.0 